--- Tribler/Test/Community/Multichain/StatisticsDispaly/__init__.py
@@ -1,3 +1,0 @@
-"""
-This package contains tests for the Statistics display in Tribler.
-"""
