+++ Tribler/Test/Community/Multichain/statistics/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the trust display in Tribler.
+"""
