--- Tribler/community/multichain/StatisticsDisplay/__init__.py
@@ -1,3 +1,0 @@
-"""
-The Statistics Display is a display for the trust statistics in the local network.
-"""
