+++ Tribler/community/multichain/statistics/__init__.py
@@ -1,0 +1,3 @@
+"""
+The statistics package provides the data for the trust display.
+"""
